v1.6: correctif de manière à mieux tracer les exceptions se produisant dans les services

+++ src/main/java/fr/diginamic/composants/ui/Selectable.java
@@ -10,5 +10,5 @@ public interface Selectable {
 	 * C'est cet id qui sera transmis lorsque l'utilisateur aura fait sa sélection.
 	 * @return Long
 	 */
-	Long getId();
+	int getId();
 }
